Move constraints to contracts folder, per review
