+++ core/dbt/common/contracts/constraints.py
