+++ core/dbt/parser/schemas.py
@@ -5,7 +5,7 @@ from abc import ABCMeta, abstractmethod
 from typing import Any, Callable, Dict, Generic, Iterable, List, Optional, Type, TypeVar
 from dataclasses import dataclass, field
 
-from dbt.common.constraints import ConstraintType, ModelLevelConstraint
+from dbt.common.contracts.constraints import ConstraintType, ModelLevelConstraint
 from dbt.common.dataclass_schema import ValidationError, dbtClassMixin
 
 from dbt.clients.yaml_helper import load_yaml_text
